requirements.txt
# cad_draft_reviewing/requirements.txt

torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
pymupdf==1.21.1
easyocr==1.4
torch-geometric==2.3.0
neo4j==5.5.0
streamlit==1.25.0
pandas==2.0.1
scikit-learn==1.3.0